Moving file to detail package.
