+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/home/FeaturedCollections.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.jordangellatly.hermosahappyhour.model.*
+import com.jordangellatly.hermosahappyhour.ui.detail.HappyHourSpecialsCollection
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
 import com.jordangellatly.hermosahappyhour.ui.utils.mirroringIcon
